finish moving files to proper directories
