+++ SplineMiner/UI/DebugTools/DebugPanel.cs
@@ -1,10 +1,12 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
+using SplineMiner.Core.Services;
+using SplineMiner.Game.Items.Tools;
 using System;
 using System.Collections.Generic;
 
-namespace SplineMiner.UI
+namespace SplineMiner.UI.DebugTools
 {
     /// <summary>
     /// Manages debug information display and development tools.
